# docs/swagger.yaml
basePath: /
definitions:
...
  goals.Goals:
    properties:
      calcium_mg:
        $ref: '#/definitions/goals.Range'
      carbs_g:
        $ref: '#/definitions/goals.Range'
      created_at:
        type: string
      fat_g:
        $ref: '#/definitions/goals.Range'
      fiber_g:
        $ref: '#/definitions/goals.Range'
      iron_mg:
        $ref: '#/definitions/goals.Range'
      kcal:
        $ref: '#/definitions/goals.Range'
      magnesium_mg:
        $ref: '#/definitions/goals.Range'
      potassium_mg:
        $ref: '#/definitions/goals.Range'
      protein_g:
        $ref: '#/definitions/goals.Range'
      salt_g:
        $ref: '#/definitions/goals.Range'
      sugar_g:
        $ref: '#/definitions/goals.Range'
      updated_at:
        type: string
      vitamin_b12_mcg:
        $ref: '#/definitions/goals.Range'
      vitamin_c_mg:
        $ref: '#/definitions/goals.Range'
      vitamin_d_mcg:
        $ref: '#/definitions/goals.Range'
      zinc_mg:
        $ref: '#/definitions/goals.Range'
    type: object
  goals.Range:
    properties:
      max:
        type: number
      min:
        type: number
    type: object
...
  raceprep.AppliedEntry:
    properties:
      carbs_g_min:
        type: number
      created:
        type: boolean
      date:
        type: string
    type: object
  raceprep.ApplyResponse:
    properties:
      applied:
        items:
          $ref: '#/definitions/raceprep.AppliedEntry'
        type: array
      body_weight_kg:
        type: number
      params:
        $ref: '#/definitions/raceprep.EchoedParams'
      race_date:
        type: string
      schedule:
        items:
          $ref: '#/definitions/raceprep.CarbLoadEntry'
        type: array
    type: object
  raceprep.CarbLoadEntry:
    properties:
      date:
        description: YYYY-MM-DD
        type: string
      days_before:
        description: 0 for race day, N for N days before
        type: integer
      rationale:
        type: string
      target_carbs_g:
        description: rounded to 1 dp
        type: number
    type: object
  raceprep.CarbLoadSchedule:
    properties:
      body_weight_kg:
        type: number
      params:
        $ref: '#/definitions/raceprep.EchoedParams'
      race_date:
        type: string
      schedule:
        items:
          $ref: '#/definitions/raceprep.CarbLoadEntry'
        type: array
    type: object
  raceprep.EchoedParams:
    properties:
      carbs_per_kg_per_day:
        type: number
      days_before:
        type: integer
      race_day_carbs_per_kg:
        type: number
    type: object
  raceprep.FuelRecommendation:
    properties:
      inputs:
        $ref: '#/definitions/raceprep.RecommendInputs'
      intra_workout:
        $ref: '#/definitions/raceprep.IntraWorkout'
      notes:
        items:
          type: string
        type: array
      post_workout:
        $ref: '#/definitions/raceprep.PostWorkout'
      pre_workout:
        $ref: '#/definitions/raceprep.PreWorkout'
    type: object
  raceprep.IntraWorkout:
    properties:
      applicable:
        type: boolean
      carbs_g_per_hour:
        type: number
      carbs_g_total:
        type: number
      fluid_ml_per_hour:
        type: number
      rationale:
        type: string
      sodium_mg_per_hour:
        type: number
    type: object
  raceprep.PostWorkout:
    properties:
      carbs_g:
        type: number
      protein_g:
        type: number
      rationale:
        type: string
      window_minutes_after:
        items:
          type: integer
        type: array
    type: object
  raceprep.PreWorkout:
    properties:
      carbs_g:
        type: number
      carbs_g_per_kg:
        type: number
      rationale:
        type: string
      window_minutes_before:
        items:
          type: integer
        type: array
    type: object
  raceprep.RecommendInputs:
    properties:
      body_weight_kg:
        type: number
      body_weight_source:
        type: string
      duration_min:
        type: integer
      intensity_zone:
        type: integer
      sport:
        type: string
      workout_id:
        type: string
    type: object
  raceprep.applyRequestBody:
    properties:
      body_weight_kg:
        type: number
      carbs_per_kg_per_day:
        type: number
      days_before:
        type: integer
      race_date:
        type: string
      race_day_carbs_per_kg:
        type: number
    type: object
...
info:
  contact: {}
  description: Kazper — personal endurance-fueling and training-coach REST API. All
    endpoints under /products, /meals, and /summary require a bearer token.
  title: Kazper
  version: 0.1.0
paths:
...
  /goals/overrides/{date}:
    delete:
      parameters:
      - description: Override date in YYYY-MM-DD
        in: path
        name: date
        required: true
        type: string
      responses:
        "204":
          description: no content
        "400":
          description: date_invalid
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: override_not_found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a daily goal override
      tags:
      - goals
    get:
      parameters:
      - description: Override date in YYYY-MM-DD
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{\"goals\": Goals}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: date_invalid
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: override_not_found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a daily goal override for a specific date
      tags:
      - goals
    put:
      consumes:
      - application/json
      description: 'Full-replace semantics: absent fields are stored as NULL. Same
        validation as PUT /goals. `Idempotency-Key` is NOT accepted on PUT — supplying
        it returns `400 idempotency_unsupported_for_put`.'
      parameters:
      - description: Override date in YYYY-MM-DD
        in: path
        name: date
        required: true
        type: string
      - description: Goals payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/goals.Goals'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"goals\": Goals}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: date_invalid | goal_value_invalid | goal_range_invalid | invalid_json
            | idempotency_unsupported_for_put
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upsert a daily goal override
      tags:
      - goals
...
  /race-prep/carb-load:
    get:
      description: 'Stateless, deterministic carb-load math: given a race date and
        a body weight, returns daily carbohydrate targets for the load window and
        race morning. No persistence. The natural follow-up is to translate each entry
        into a goal override via `PUT /goals/overrides/{date}`.'
      parameters:
      - description: Race date YYYY-MM-DD (must be today or later in DEFAULT_USER_TZ)
        in: query
        name: race_date
        required: true
        type: string
      - description: Athlete body weight in kilograms, 30..200
        in: query
        name: body_weight_kg
        required: true
        type: number
      - description: Carb-load days before race day, 0..7 (default 3)
        in: query
        name: days_before
        type: integer
      - description: Load-day multiplier, 1..20 g/kg (default 10)
        in: query
        name: carbs_per_kg_per_day
        type: number
      - description: Race-morning multiplier, 0..10 g/kg (default 2)
        in: query
        name: race_day_carbs_per_kg
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/raceprep.CarbLoadSchedule'
        "400":
          description: race_date_required | body_weight_kg_required | race_date_invalid
            | race_date_in_past | body_weight_kg_invalid | days_before_invalid | carbs_per_kg_per_day_invalid
            | race_day_carbs_per_kg_invalid
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Compute a carb-load schedule for a race
      tags:
      - race-prep
  /race-prep/carb-load/apply:
    post:
      consumes:
      - application/json
      description: Same inputs as `GET /race-prep/carb-load`. Computes the schedule,
        then in a single transaction writes the per-day carb target into the `daily_goal_overrides`
        row for each schedule date — merging into existing overrides (preserving non-carb
        fields) or creating new rows. Returns the schedule plus a per-date `applied`
        outcome (`{date, carbs_g_min, created}`). On any per-date failure the whole
        transaction rolls back and zero overrides are persisted.
      parameters:
      - description: Carb-load parameters (same as the GET endpoint)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/raceprep.applyRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/raceprep.ApplyResponse'
        "400":
          description: race_date_required | body_weight_kg_required | race_date_invalid
            | race_date_in_past | body_weight_kg_invalid | days_before_invalid | carbs_per_kg_per_day_invalid
            | race_day_carbs_per_kg_invalid | invalid_json
          schema:
            additionalProperties: true
            type: object
        "500":
          description: apply_failed
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Compute AND apply a carb-load schedule into per-date goal overrides
      tags:
      - race-prep
  /race-prep/recommend-workout-fuel:
    get:
      description: Stateless literature-grounded fueling recommendation for one training
        or race session. Accepts EITHER `workout_id` (pulls sport/duration/intensity
        from the row, deriving intensity zone from `tss` when present) OR the explicit
        triplet `sport`+`duration_min`+`intensity_zone`. Body weight resolved via
        the four-tier rule (explicit override > rolling 7d stored mean > most-recent
        stored entry > 400). Reuses the 0.3 g/kg MPS threshold from `protein_distribution`
        for the post-workout protein recommendation.
      parameters:
      - description: Workout UUID — pulls sport/duration/intensity from the row
        in: query
        name: workout_id
        type: string
      - description: Sport (bike|run|swim|strength|other). Required in explicit mode.
        in: query
        name: sport
        type: string
      - description: Duration in minutes (>0). Required in explicit mode.
        in: query
        name: duration_min
        type: integer
      - description: Zone 1–5. Required in explicit mode.
        in: query
        name: intensity_zone
        type: integer
      - description: Explicit body-weight override (>0). Otherwise resolved from stored
          entries.
        in: query
        name: body_weight_kg
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/raceprep.FuelRecommendation'
        "400":
          description: input_required | input_conflict | sport_required | duration_min_required
            | intensity_zone_required | sport_invalid | duration_min_invalid | intensity_zone_invalid
            | body_weight_kg_invalid | weight_data_missing | workout_id_invalid
          schema:
            additionalProperties: true
            type: object
        "404":
          description: workout_not_found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Pre/intra/post fueling recommendation for a single session
      tags:
      - race-prep
...
securityDefinitions:
  BearerAuth:
    description: Static bearer token. Use the format `Bearer <token>` with either
      MOBILE_API_TOKEN or AGENT_API_TOKEN.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"